Update main page layout
- Fixes and adjustments for the carousel
- Main page text intro is now a separate component
