+++ src/components/common/MainViewCarousel.js
@@ -1,10 +1,11 @@
 import React from 'react'
 import { Carousel } from 'react-bootstrap'
+
 import styles from './MainViewCarousel.module.sass'
 
 const MainViewCarousel = () => {
 	return (
-		<Carousel className={styles.carouselContainer} slide>
+		<Carousel className={styles.carouselContainer} interval={20000}>
 			<Carousel.Item>
 				<div className={styles.carouselOverlay}></div>
 				<img
@@ -12,9 +13,8 @@ const MainViewCarousel = () => {
 					src="img/carousel/arts-n-crafts.jpg"
 					alt="First slide"
 				/>
-				<Carousel.Caption>
-					<h3>Напрямки</h3>
-					<br />
+				<Carousel.Caption className={styles.caption}>
+					<h3 className={styles.slideTitle}>Напрямки</h3>
 					<p>Образотворче мистецтво, музика, театр</p>
 				</Carousel.Caption>
 			</Carousel.Item>
@@ -26,9 +26,8 @@ const MainViewCarousel = () => {
 					src="img/carousel/paper-boats.jpg"
 					alt="Third slide"
 				/>
-				<Carousel.Caption>
-					<h3>Вчителі</h3>
-					<br />
+				<Carousel.Caption className={styles.caption}>
+					<h3 className={styles.slideTitle}>Вчителі</h3>
 					<p>Наші вчителі найдосвідченіші! )</p>
 				</Carousel.Caption>
 			</Carousel.Item>
@@ -40,9 +39,8 @@ const MainViewCarousel = () => {
 					src="img/carousel/commerce.jpg"
 					alt="Third slide"
 				/>
-				<Carousel.Caption>
-					<h3>Навчання</h3>
-					<br />
+				<Carousel.Caption className={styles.caption}>
+					<h3 className={styles.slideTitle}>Навчання</h3>
 					<p>Заповнити заявку ви можете у нас на сайті</p>
 				</Carousel.Caption>
 			</Carousel.Item>
