--- src/components/common/layout/Parallax.module.sass
@@ -1,3 +1,0 @@
-
-.parallaxSection
-	margin: 0rem 0
