+++ src/components/views/public/MainPage.module.sass
@@ -1,30 +1,6 @@
 @import '../../../sass/variables'
 
 .parallax
-	margin: 6rem 0
+	margin: 8rem 0
 	@media screen and (min-width: $medium)
-		margin: 10rem 0
-
-.logoImg
-	max-width: 400px
-	margin: 4rem auto 0
-
-.mainTextItem
-	padding: 4rem 0
-	@media screen and (min-width: $medium)
-		font-size: 1.1rem
-	p
-		text-indent: 1.4em
-		line-height: 1.5
-		margin-top: 2rem
-		max-width: 500px
-		margin: auto
-	&:last-child
-		padding: 0
-
-.animIconContainer,
-.ctaContainer
-	text-align: center
-
-.animIconContainer
-	margin-bottom: 3rem
+		margin: 11rem 0
