+++ src/sass/_text.sass
@@ -1,29 +1,6 @@
 @import 'mixins.sass'
 @import 'palette.scss'
 
-.carousel-caption
-	text-align: left
-	line-height: 1.1
-	padding-top: 0
-
-.carousel-caption h3
-	margin-bottom: -1rem
-	color: white
-	+desktop(96dpi, 600px)
-		font-size: 200%
-	+desktop(96dpi, 1024px)
-		font-size: 400%
-	+handheld(2, 1440px)
-		font-size: 300%
-
-.carousel-caption p
-	+desktop(96dpi, 600px)
-		font-size: 150%
-	+desktop(96dpi, 1024px)
-		font-size: 200%
-	+handheld(2, 1440px)
-		font-size: 170%
-
 .main-title
 	font-size: 150%
 	text-align: center
